feat: initial shape
